+++ src/containment.rs
@@ -38,6 +38,6 @@ impl ContainmentGraph {
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct AABB {
-    pub min: Point3<f64>,
-    pub max: Point3<f64>,
+    pub min: Point3<f32>,
+    pub max: Point3<f32>,
 }
